+++ src/Addons/34thPRC_ArmourStandard/data/marines2/woodland/config_weapons.hpp
@@ -482,7 +482,7 @@ class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Red : 34
 };
 class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
 {
-	displayName="[34th] Marine Cross Branch BDU Woodland (Light)";
+	displayName="[34th] BDU Woodland (Light)";
 	author="Ambrose";
 	class ItemInfo: ItemInfo
 	{
@@ -498,7 +498,7 @@ class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light: 34thPRC_ArmourStan
 };
 class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
 {
-	displayName="[34th] Marine Cross Branch BDU Woodland (Medium)";
+	displayName="[34th] BDU Woodland (Medium)";
 	author="Ambrose";
 	class ItemInfo: ItemInfo
 	{
@@ -514,7 +514,7 @@ class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium: 34thPRC_ArmourSta
 };
 class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
 {
-	displayName="[34th] Marine Cross Branch BDU Woodland (Heavy)";
+	displayName="[34th] BDU Woodland (Heavy)";
 	author="Ambrose";
 	class ItemInfo: ItemInfo
 	{
@@ -530,7 +530,7 @@ class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy: 34thPRC_ArmourStan
 };
 class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
 {
-	displayName="[34th] Marine Cross Branch BDU Woodland (Shoulderless)";
+	displayName="[34th] BDU Woodland (Shoulderless)";
 	author="Ambrose";
 	class ItemInfo: ItemInfo
 	{
@@ -546,7 +546,7 @@ class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_NoShoulder: 34thPRC_Armou
 };
 class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
 {
-	displayName="[34th] Marine Cross Branch BDU Woodland (Light) (No Collar)";
+	displayName="[34th] BDU Woodland (Light) (No Collar)";
 	author="Ambrose";
 	class ItemInfo: ItemInfo
 	{
@@ -562,7 +562,7 @@ class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light_NoCollar: 34thPRC_A
 };
 class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
 {
-	displayName="[34th] Marine Cross Branch BDU Woodland (Medium) (No Collar)";
+	displayName="[34th] BDU Woodland (Medium) (No Collar)";
 	author="Ambrose";
 	class ItemInfo: ItemInfo
 	{
@@ -578,7 +578,7 @@ class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium_NoCollar: 34thPRC_
 };
 class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
 {
-	displayName="[34th] Marine Cross Branch BDU Woodland (Heavy) (No Collar)";
+	displayName="[34th] BDU Woodland (Heavy) (No Collar)";
 	author="Ambrose";
 	class ItemInfo: ItemInfo
 	{