+++ src/changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - All Spartan shoulders have gunbags
 - Added black hot and white hot thermal modes to UNSC Marksma's Visual Interface and all MVI tagged NVGs
+- Shortened all Marine Cross Branch BDU uniform display names to BDU and fixed name on Urban Blue 'Jade' Medium custom armor, was 'Demo Player'
 
 ## 0.20.0
 ### Added